third_party/WebKit/LayoutTests/platform/mac/fast/events/platform-wheelevent-paging-x-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is 0
PASS event.wheelDeltaX is window.givenScrollLeft * -3
PASS event.wheelDelta is window.givenScrollLeft * -3
PASS document.scrollingElement.scrollTop is window.expectedScrollTop
FAIL document.scrollingElement.scrollLeft should be 700. Was 760.
PASS successfullyParsed is true

TEST COMPLETE
